+++ .gitignore
@@ -58,3 +58,5 @@ preload.js
 preload.js.map
 main.js
 main.js.map
+export-marker-task.js
+export-marker-task.js.map
