+++ package.json
@@ -8,6 +8,7 @@
     "name": "OPENCOLLAB",
     "email": "sakai@opencollab.co.za"
   },
+  "sideEffects": false,
   "repository": {
     "type": "git",
     "url": "https://github.com/efundi/PDF-Marker.git"
@@ -24,16 +25,17 @@
     "postinstall": "electron-builder install-app-deps",
     "start": "npm-run-all -p ng:serve electron:serve",
     "debug": "npm-run-all -p ng:serve electron:debug",
-    "build": "npm run electron:webpack && ng build",
-    "build:dev": "npm run build --",
-    "build:prod": "npm run build -- -c production",
+    "build": "npm run build:dev",
+    "build:dev": "npm run electron:webpack:dev && npm run ng:build:dev",
+    "build:prod": "npm run electron:webpack:prod && npm run ng:build:prod",
     "deploy": "npm run build:prod && electron-builder build --win --publish always",
-    "ng:build": "ng build --configuration production",
+    "ng:build:prod": "ng build --configuration production",
+    "ng:build:dev": "ng build",
     "ng:serve": "ng serve",
-    "ng:serve:web": "&& ng serve -o",
-    "electron:webpack": "webpack --config webpack.electron.config.js --progress",
-    "electron:serve": "wait-on http-get://localhost:4200/ && npm run electron:webpack && electron . --inspect=5858 --serve",
-    "electron:debug": "wait-on http-get://localhost:4200/ && npm run electron:webpack && electron . --inspect-brk=5858 --serve",
+    "electron:webpack:dev": "webpack --config webpack.electron.config.js --progress",
+    "electron:webpack:prod": "webpack --config webpack.electron.config.js --progress --env production=true",
+    "electron:serve": "wait-on http-get://localhost:4200/ && npm run electron:webpack:dev && electron . --inspect=5858 --serve",
+    "electron:debug": "wait-on http-get://localhost:4200/ && npm run electron:webpack:dev && electron . --inspect-brk=5858 --serve",
     "electron:local": "npm run build:prod && electron .",
     "electron:linux": "npm run build:prod && electron-builder build --linux",
     "electron:windows": "npm run build:prod && electron-builder build --windows",
