+++ src-electron/ipc/assignment.handler.ts
@@ -67,9 +67,7 @@ import {WorkerPool} from '../worker_pool';
 import {cpus} from 'os';
 
 const zipDir = require('zip-dir');
-
 const csvtojson = require('csvtojson');
-
 
 export function getAssignments(): Promise<Workspace[]> {
   return loadWorkspaces();
@@ -568,6 +566,8 @@ export function writeAssignmentSettingsAt(
 
   return writeFile(assignmentAbsolutePath + sep + SETTING_FILE, buffer).then(() => {
     return assignmentSettings;
+  }, () => {
+    return Promise.reject('Failed to save assignment settings!');
   });
 }
 
@@ -810,8 +810,6 @@ export function finalizeAssignment(event: IpcMainInvokeEvent, workspaceFolder: s
   }
 }
 
-
-
 function cleanupTemp(tmpDir: string) {
   try {
     if (tmpDir) {
@@ -832,66 +830,61 @@ export function exportAssignment(exportAssignmentsRequest: ExportAssignmentsRequ
     return exportForModeration(exportAssignmentsRequest);
   }
 
-  return new Promise<void>((resolve, reject) => {
 
-  });
+  const tempDirectory = mkdtempSync(join(os.tmpdir(), 'pdfm-'));
+  const exportTempDirectory = tempDirectory + sep + exportAssignmentsRequest.assignmentName;
 
+  return Promise.all([
+    getAssignmentSettingsFor(exportAssignmentsRequest.workspaceFolder, exportAssignmentsRequest.assignmentName),
+    getAssignmentDirectoryAbsolutePath(exportAssignmentsRequest.workspaceFolder, exportAssignmentsRequest.assignmentName),
+    mkdir(exportTempDirectory)
+  ])
+    .then(([assignmentSettings, originalAssignmentDirectory]) => {
+      let exportSubmissions: Submission[] = assignmentSettings.submissions;
 
-  //
-  // const tempDirectory = mkdtempSync(join(os.tmpdir(), 'pdfm-'));
-  // const exportTempDirectory = tempDirectory + sep + exportAssignmentsRequest.assignmentName;
-  //
-  // return Promise.all([
-  //   getAssignmentSettingsFor(exportAssignmentsRequest.workspaceFolder, exportAssignmentsRequest.assignmentName),
-  //   getAssignmentDirectoryAbsolutePath(exportAssignmentsRequest.workspaceFolder, exportAssignmentsRequest.assignmentName),
-  //   mkdir(exportTempDirectory)
-  // ])
-  //   .then(([assignmentSettings, originalAssignmentDirectory]) => {
-  //     let exportSubmissions: Submission[] = assignmentSettings.submissions;
-  //
-  //     if (!isNil(exportAssignmentsRequest.studentIds)) {
-  //       // If a list of student ids was supplied we'll filter only those
-  //       exportSubmissions = filter(assignmentSettings.submissions, (submission) => {
-  //         return exportAssignmentsRequest.studentIds.indexOf(submission.studentId) >= 0;
-  //       });
-  //     }
-  //
-  //     // Copy all the submission files
-  //     const promises: Promise<any>[] = exportSubmissions.map((submission) => {
-  //       return copy(
-  //         originalAssignmentDirectory + sep + submission.directoryName,
-  //         exportTempDirectory + sep + submission.directoryName,
-  //         {
-  //           recursive: true,
-  //           //   filter: (src) => {
-  //           //     return !src.endsWith(MARK_FILE);
-  //           //   }
-  //         }
-  //       );
-  //     });
-  //     return Promise.all(promises)
-  //       .then(() => {
-  //         // We need to create a settings file
-  //         const exportSettings = cloneDeep(assignmentSettings);
-  //         if (!isNil(exportAssignmentsRequest.studentIds)) {
-  //           // Filter out submissions if required
-  //           exportSettings.submissions = exportSettings.submissions.filter((submission) => {
-  //             return exportAssignmentsRequest.studentIds.indexOf(submission.studentId) >= 0;
-  //           });
-  //         }
-  //         return writeAssignmentSettingsAt(exportSettings, exportTempDirectory);
-  //       })
-  //       .then(() => {
-  //         return zipDir(tempDirectory);
-  //       })
-  //       .then((buffer) => {
-  //         cleanupTemp(tempDirectory);
-  //         return buffer;
-  //       }, (error) => {
-  //         cleanupTemp(tempDirectory);
-  //         return Promise.reject(error.message);
-  //       });
-  //   });
+      if (!isNil(exportAssignmentsRequest.studentIds)) {
+        // If a list of student ids was supplied we'll filter only those
+        exportSubmissions = filter(assignmentSettings.submissions, (submission) => {
+          return exportAssignmentsRequest.studentIds.indexOf(submission.studentId) >= 0;
+        });
+      }
+
+      // Copy all the submission files
+      const promises: Promise<any>[] = exportSubmissions.map((submission) => {
+        return copy(
+          originalAssignmentDirectory + sep + submission.directoryName,
+          exportTempDirectory + sep + submission.directoryName,
+          {
+            recursive: true,
+            //   filter: (src) => {
+            //     return !src.endsWith(MARK_FILE);
+            //   }
+          }
+        );
+      });
+      return Promise.all(promises)
+        .then(() => {
+          // We need to create a settings file
+          const exportSettings = cloneDeep(assignmentSettings);
+          if (!isNil(exportAssignmentsRequest.studentIds)) {
+            // Filter out submissions if required
+            exportSettings.submissions = exportSettings.submissions.filter((submission) => {
+              return exportAssignmentsRequest.studentIds.indexOf(submission.studentId) >= 0;
+            });
+          }
+          return writeAssignmentSettingsAt(exportSettings, exportTempDirectory);
+        })
+        .then(() => {
+          return zipDir(tempDirectory);
+        })
+        .then((buffer) => {
+          cleanupTemp(tempDirectory);
+          return buffer;
+        }, (error) => {
+          cleanupTemp(tempDirectory);
+          return Promise.reject(error.message);
+        });
+    });
 }
 
 
@@ -1022,31 +1015,22 @@ export function generateAllocationZipFiles(event: IpcMainInvokeEvent,
       }, {});
 
 
-      const pool = new WorkerPool(cpus().length);
-
-
-
+      let pool = new WorkerPool(cpus().length);
       const promises: Promise<any>[] = map(Object.keys(allocationSubmissionsMap), (markerEmail) => {
-        return new Promise<void>((resolve, reject) => {
-          pool.runTask({
-            format: ExportFormat.PDFM,
-            studentIds: allocationSubmissionsMap[markerEmail],
-            workspaceFolder: workspaceName,
-            assignmentName: assignmentName,
-            exportPath,
-            markerEmail
-          }, (err, result) => {
-            console.log("task done");
-            console.log(err, result);
-            resolve();
-          });
+        return pool.queueTask({
+          format: ExportFormat.PDFM,
+          studentIds: allocationSubmissionsMap[markerEmail],
+          workspaceFolder: workspaceName,
+          assignmentName: assignmentName,
+          exportPath,
+          markerEmail
         });
       });
 
       return Promise.all(promises)
         .then(() => {
-          console.log("All done");
           pool.close();
+          pool = null;
         });
     }).then(() => Promise.resolve()); // End with noop to make it return no value
 }
